requirements.txt
pygame-ce>=2.5.0
neat-python
